Binding in contacts list: use "M" to send a message to contact
The idea of bindings is that you can get to any screen from any screen with
common keys: "s" for the status, "m" for the messages, "c" for contacts,
"l" for the calllog.

Where it makes sense, "r" means reply (with message if you're viewing a
message or place a call if you looking in the call log), "C" -- call the
peer, "M" -- message the peer. A special exception -- "RET" and left mouse
click calls current contact in the contact list.